int -> unsiged long

+++ tests/testlevenshtein.cc
@@ -4,9 +4,9 @@
 
 TEST_CASE("levenshtein")
 {
-  const std::string a        = "kitten";
-  const std::string b        = "sitting";
-  cons int          distance = 3;
+  const std::string  a        = "kitten";
+  const std::string  b        = "sitting";
+  cons unsigned long distance = 3;
 
   CHECK(LevenshteinDistance(a, b) == distance);
   CHECK(LevenshteinDistance(b, a) == distance);
